report on progress - 1 remain

+++ api/src/main/java/org/openmrs/module/PSI/dto/AUHCComprehensiveReport.java
@@ -3,18 +3,18 @@ package org.openmrs.module.PSI.dto;
 public class AUHCComprehensiveReport {
 	private String service_code;
 	private String service_name;
-	private Long service_contact_static;
-	private Long service_contact_satellite;
-	private Long service_contact_csp;
-	private Long service_total;
-	private Double revenue_static;
-	private Double revenue_satellite;
-	private Double revenue_csp;
-	private Double revenue_total;
-	private Double discount_static;
-	private Double discount_satellite;
-	private Double discount_csp;
-	private Double discount_total;
+	private float service_contact_static;
+	private float service_contact_satellite;
+	private float service_contact_csp;
+	private float service_total;
+	private float revenue_static;
+	private float revenue_satellite;
+	private float revenue_csp;
+	private float revenue_total;
+	private float discount_static;
+	private float discount_satellite;
+	private float discount_csp;
+	private float discount_total;
 	public String getService_code() {
 		return service_code;
 	}
@@ -27,78 +27,79 @@ public class AUHCComprehensiveReport {
 	public void setService_name(String service_name) {
 		this.service_name = service_name;
 	}
-	public Long getService_contact_static() {
+	public float getService_contact_static() {
 		return service_contact_static;
 	}
-	public void setService_contact_static(Long service_contact_static) {
+	public void setService_contact_static(float service_contact_static) {
 		this.service_contact_static = service_contact_static;
 	}
-	public Long getService_contact_satellite() {
+	public float getService_contact_satellite() {
 		return service_contact_satellite;
 	}
-	public void setService_contact_satellite(Long service_contact_satellite) {
+	public void setService_contact_satellite(float service_contact_satellite) {
 		this.service_contact_satellite = service_contact_satellite;
 	}
-	public Long getService_contact_csp() {
+	public float getService_contact_csp() {
 		return service_contact_csp;
 	}
-	public void setService_contact_csp(Long service_contact_csp) {
+	public void setService_contact_csp(float service_contact_csp) {
 		this.service_contact_csp = service_contact_csp;
 	}
-	public Long getService_total() {
+	public float getService_total() {
 		return service_total;
 	}
-	public void setService_total(Long service_total) {
+	public void setService_total(float service_total) {
 		this.service_total = service_total;
 	}
-	public Double getRevenue_static() {
+	public float getRevenue_static() {
 		return revenue_static;
 	}
-	public void setRevenue_static(Double revenue_static) {
+	public void setRevenue_static(float revenue_static) {
 		this.revenue_static = revenue_static;
 	}
-	public Double getRevenue_satellite() {
+	public float getRevenue_satellite() {
 		return revenue_satellite;
 	}
-	public void setRevenue_satellite(Double revenue_satellite) {
+	public void setRevenue_satellite(float revenue_satellite) {
 		this.revenue_satellite = revenue_satellite;
 	}
-	public Double getRevenue_csp() {
+	public float getRevenue_csp() {
 		return revenue_csp;
 	}
-	public void setRevenue_csp(Double revenue_csp) {
+	public void setRevenue_csp(float revenue_csp) {
 		this.revenue_csp = revenue_csp;
 	}
-	public Double getRevenue_total() {
+	public float getRevenue_total() {
 		return revenue_total;
 	}
-	public void setRevenue_total(Double revenue_total) {
+	public void setRevenue_total(float revenue_total) {
 		this.revenue_total = revenue_total;
 	}
-	public Double getDiscount_static() {
+	public float getDiscount_static() {
 		return discount_static;
 	}
-	public void setDiscount_static(Double discount_static) {
+	public void setDiscount_static(float discount_static) {
 		this.discount_static = discount_static;
 	}
-	public Double getDiscount_satellite() {
+	public float getDiscount_satellite() {
 		return discount_satellite;
 	}
-	public void setDiscount_satellite(Double discount_satellite) {
+	public void setDiscount_satellite(float discount_satellite) {
 		this.discount_satellite = discount_satellite;
 	}
-	public Double getDiscount_csp() {
+	public float getDiscount_csp() {
 		return discount_csp;
 	}
-	public void setDiscount_csp(Double discount_csp) {
+	public void setDiscount_csp(float discount_csp) {
 		this.discount_csp = discount_csp;
 	}
-	public Double getDiscount_total() {
+	public float getDiscount_total() {
 		return discount_total;
 	}
-	public void setDiscount_total(Double discount_total) {
+	public void setDiscount_total(float discount_total) {
 		this.discount_total = discount_total;
 	}
+	
 	
 	
 }
